requirements.txt
aiogram>=3.20.0
fastapi>=0.110.0
groq>=1.0.0
httpx>=0.27.0
ollama>=0.4.0
python-dotenv>=1.0.0
requests>=2.31.0
rich>=13.0.0
uvicorn>=0.29.0
